requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
click==8.1.7
Django==4.2.15
exceptiongroup==1.2.2
fastapi==0.101.1
h11==0.14.0
httptools==0.6.1
idna==3.7
image==1.5.33
Pillow==9.3.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.22.0
uvloop==0.19.0
watchfiles==0.23.0
websockets==12.0
